+++ app/src/main/java/com/mywaytec/myway/model/bean/VehicleTrackBean.java
@@ -34,7 +34,7 @@ public class VehicleTrackBean extends BaseInfo{
         private String id;
         private String deviceId;
         private String deviceType;
-        private int frameTime;
+        private long frameTime;
         private int index;
         private LocBean loc;
 
@@ -62,11 +62,11 @@ public class VehicleTrackBean extends BaseInfo{
             this.deviceType = deviceType;
         }
 
-        public int getFrameTime() {
+        public long getFrameTime() {
             return frameTime;
         }
 
-        public void setFrameTime(int frameTime) {
+        public void setFrameTime(long frameTime) {
             this.frameTime = frameTime;
         }
